Add Input class for handling key input; Update main.py
Window class now initializes an Input class that can be accessed using
the Window.input attribute. Calling Input.key_events will return a list
of tuples with all key codes and their state (pressed or released) that
happened since the last time Input.key_events was called. This method
should be used when a key press is expected to be pressed once to open
menus for example.

Calling Input.key_pressed will return True or False whether the given
key is currently being pressed or not. Currently is important here, as
it will not return True if a key was being pressed before this call.
This method should be used for movement for examplle.